update and authorized

+++ src/main/java/com/example/qairline/controller/AnnouncementController.java
@@ -7,6 +7,7 @@ import com.example.qairline.dto.response.AircraftResponse;
 import com.example.qairline.dto.response.AnnouncementResponse;
 import com.example.qairline.dto.response.FlightResponse;
 import com.example.qairline.service.AnnouncementService;
+import com.example.qairline.util.SecurityUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Controller;
@@ -19,23 +20,36 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 public class AnnouncementController {
     @Autowired
     private AnnouncementService announcementService;
+    @Autowired
+    private SecurityUtil securityUtil;
+    @Autowired
+    AuthController authController;
 
     @GetMapping("api/announcements")
     @ResponseBody
-    public Page<AnnouncementResponse> getAllAircrafts(@RequestParam int page, @RequestParam int size) {
-        return announcementService.getAllAnnouncementPaging(page, size);
+    public Page<AnnouncementResponse> getAllAircrafts(@RequestHeader("Authorization") String authorizationHeader, @RequestParam int page, @RequestParam int size) {
+        if (isAdmin() && !authController.isTokenBlacklisted(authorizationHeader)) {
+            return announcementService.getAllAnnouncementPaging(page, size);
+        }
+        return null;
     }
 
     @GetMapping("api/announcementsSearch")
     @ResponseBody
-    public Page<AnnouncementResponse> getUsersSearch(@RequestParam String keyword, @RequestParam int page, @RequestParam int size) {
-        return announcementService.searchAnnouncements(keyword, page, size);
+    public Page<AnnouncementResponse> getUsersSearch(@RequestHeader("Authorization") String authorizationHeader, @RequestParam String keyword, @RequestParam int page, @RequestParam int size) {
+        if (isAdmin() && !authController.isTokenBlacklisted(authorizationHeader)) {
+            return announcementService.searchAnnouncements(keyword, page, size);
+        }
+        return null;
     }
 
     @GetMapping("/api/announcements/{announcementId}")
     @ResponseBody
-    public AnnouncementResponse getFlightById(@PathVariable Long announcementId) {
-        return announcementService.getAnnouncementById(announcementId);
+    public AnnouncementResponse getFlightById(@RequestHeader("Authorization") String authorizationHeader, @PathVariable Long announcementId) {
+        if (isAdmin() && !authController.isTokenBlacklisted(authorizationHeader)) {
+            return announcementService.getAnnouncementById(announcementId);
+        }
+        return null;
     }
 
     @GetMapping("/admin/announcements")
@@ -45,26 +59,46 @@ public class AnnouncementController {
         return "admin/announcements";
     }
 
-    @PostMapping("/admin/announcements")
+    @PostMapping("/admin/addannouncements")
     @ResponseBody
-    public String addAnnouncement(@ModelAttribute("addAnnouncementModelAttribute") AnnouncementRequest request, RedirectAttributes redirectAttributes) {
-        if (announcementService.addAnnouncement(request) == null) {
-            //redirectAttributes.addFlashAttribute("aircraftExisted", "Aircraft with this model already exists.");
-            return "already exists.";
-        };
-        return "success";
+    public String addAnnouncement(@RequestHeader("Authorization") String authorizationHeader, @ModelAttribute("addAnnouncementModelAttribute") AnnouncementRequest request, RedirectAttributes redirectAttributes) {
+        if (isAdmin() && !authController.isTokenBlacklisted(authorizationHeader)) {
+            if (announcementService.addAnnouncement(request) == null) {
+                //redirectAttributes.addFlashAttribute("aircraftExisted", "Aircraft with this model already exists.");
+                return "already exists.";
+            };
+            return "success";
+        }
+        return null;
     }
 
     @PostMapping("/api/announcements/{announcementId}")
     @ResponseBody
-    public AnnouncementResponse updateAnnouncements(@PathVariable Long announcementId, @RequestBody AnnouncementRequest request) {
-        return announcementService.updateAnnouncement(announcementId, request);
+    public AnnouncementResponse updateAnnouncements(@RequestHeader("Authorization") String authorizationHeader, @PathVariable Long announcementId, @RequestBody AnnouncementRequest request) {
+        if (isAdmin() && !authController.isTokenBlacklisted(authorizationHeader)) {
+            return announcementService.updateAnnouncement(announcementId, request);
+
+        }
+        return null;
     }
 
     @PostMapping("/api/deleteannouncement")
     @ResponseBody
-    public String deleteAnnouncements(@RequestParam("announcementId") Long announcementId) {
-        announcementService.deleteAnnouncement(announcementId);
-        return "success";
+    public String deleteAnnouncements(@RequestHeader("Authorization") String authorizationHeader, @RequestParam("announcementId") Long announcementId) {
+        if (isAdmin() && !authController.isTokenBlacklisted(authorizationHeader)) {
+            announcementService.deleteAnnouncement(announcementId);
+            return "success";
+        }
+        return null;
+    }
+
+    boolean isAdmin() {
+        String userName = securityUtil.getCurrentUserLogin().orElse(null);
+        System.out.println(userName);
+        // Kiểm tra role và trả về kết quả
+        if (userName != null && userName.equals("admin")) {
+            return true;
+        }
+        return false;
     }
 }
