+++ src/main/java/com/example/qairline/controller/TicketController.java
@@ -1,8 +1,11 @@
 package com.example.qairline.controller;
 
+import com.example.qairline.dto.request.FlightRequest;
 import com.example.qairline.dto.request.TicketRequest;
+import com.example.qairline.dto.response.FlightResponse;
 import com.example.qairline.dto.response.TicketResponse;
 import com.example.qairline.service.TicketService;
+import com.example.qairline.util.SecurityUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Controller;
@@ -18,30 +21,62 @@ import java.util.List;
 public class TicketController {
     @Autowired
     private TicketService ticketService;
+    @Autowired
+    private SecurityUtil securityUtil;
+    @Autowired
+    AuthController authController;
+
 
     @Transactional
     @GetMapping("/api/tickets")
     @ResponseBody
-    public Page<TicketResponse> getAllFlights(@RequestParam(defaultValue = "0") int page,
+    public Page<TicketResponse> getAllTickets(@RequestHeader("Authorization") String authorizationHeader, @RequestParam(defaultValue = "0") int page,
                                               @RequestParam(defaultValue = "5") int size) {
-        return ticketService.getTickets(page, size); // Trả về JSON danh sách flights
+        if (isAdmin() && !authController.isTokenBlacklisted(authorizationHeader)) {
+            return ticketService.getTickets(page, size);
+        }
+        return null;
     }
 
     @GetMapping("/api/getAllTickets")
     @ResponseBody
-    public List<TicketResponse> getAllFlights() {
-        return ticketService.getAllTickets();
+    public List<TicketResponse> getAllTickets(@RequestHeader("Authorization") String authorizationHeader) {
+        if (isAdmin() && !authController.isTokenBlacklisted(authorizationHeader)) {
+            return ticketService.getAllTickets();
+        }
+        return null;
     };
 
     @GetMapping("api/ticketsSearch")
     @ResponseBody
-    public Page<TicketResponse> getFlightsSearch(@RequestParam String keyword, @RequestParam int page, @RequestParam int size) {
+    public Page<TicketResponse> getTicketsSearch(@RequestHeader("Authorization") String authorizationHeader, @RequestParam String keyword, @RequestParam int page, @RequestParam int size) {
+        if (isAdmin() && !authController.isTokenBlacklisted(authorizationHeader)) {
+            return ticketService.getTicketsSearch(keyword, page, size);
+        }
+        return null;
+    }
 
-        return ticketService.getTicketsSearch(keyword, page, size);
+    @GetMapping("api/ticketsByUserisNull")
+    @ResponseBody
+    public Page<TicketResponse> getTicketsByUserisNull(@RequestHeader("Authorization") String authorizationHeader, @RequestParam int page, @RequestParam int size) {
+        if (isAdmin() && !authController.isTokenBlacklisted(authorizationHeader)) {
+            return ticketService.getTicketsByUserisNull(page, size);
+        }
+        return null;
+    }
+
+    @GetMapping("api/ticketsByUserName")
+    @ResponseBody
+    public Page<TicketResponse> getTicketsByUserName(@RequestHeader("Authorization") String authorizationHeader, @RequestParam String username, @RequestParam int page, @RequestParam int size) {
+        if (isAdmin() && !authController.isTokenBlacklisted(authorizationHeader)) {
+            return ticketService.getTicketsByUserName(username, page, size);
+        }
+        return null;
     }
 
     @GetMapping("/admin/tickets")
-    public String getFlights(Model model) {
+    public String getTickets(Model model) {
+
         model.addAttribute("ticketsCount", ticketService.getAllTickets().stream().count());
         model.addAttribute("pageCount", (ticketService.getAllTickets().stream().count() + 4) / 5);
 
@@ -50,21 +85,56 @@ public class TicketController {
 
     @PostMapping("/api/deleteticket")
     @ResponseBody
-    public String deleteFlight(@RequestParam("ticketId") Long ticketId) {
-        ticketService.deleteTicketById(ticketId);
-        return "success";
+    public String deleteTicket(@RequestHeader("Authorization") String authorizationHeader, @RequestParam("ticketId") Long ticketId) {
+        if (isAdmin() && !authController.isTokenBlacklisted(authorizationHeader)) {
+            ticketService.deleteTicketById(ticketId);
+            ticketService.updatePromotionInTicket();
+            return "success";
+        }
+        return null;
     }
 
-    @PostMapping("/admin/tickets")
+    @Transactional
+    @PostMapping("/admin/addtickets")
     @ResponseBody
-    public String addFlight(@ModelAttribute("addFlightModelAttribute") TicketRequest request, RedirectAttributes redirectAttributes) {
-        ticketService.addTicket(request);
-        return "success";
+    public String addTicket(@RequestHeader("Authorization") String authorizationHeader, @ModelAttribute("addTicketModelAttribute") TicketRequest request, RedirectAttributes redirectAttributes) {
+        if (isAdmin() && !authController.isTokenBlacklisted(authorizationHeader)) {
+            ticketService.addTicket(request);
+            ticketService.updatePromotionInTicket();
+            return "success";
+        }
+        return null;
     }
 
     @GetMapping("/api/tickets/{ticketId}")
     @ResponseBody
-    public TicketResponse getFlightById(@PathVariable Long ticketId) {
-        return ticketService.getTicketById(ticketId);
+    public TicketResponse getTicketById(@RequestHeader("Authorization") String authorizationHeader, @PathVariable Long ticketId) {
+        if (isAdmin() && !authController.isTokenBlacklisted(authorizationHeader)) {
+            return ticketService.getTicketById(ticketId);
+        }
+        return null;
+    }
+
+    @PutMapping("/api/tickets/{ticketId}")
+    @ResponseBody
+    public TicketResponse editTicket(@RequestHeader("Authorization") String authorizationHeader, @PathVariable Long ticketId, @RequestBody TicketRequest ticketRequest) {
+        if (isAdmin() && !authController.isTokenBlacklisted(authorizationHeader)) {
+            ticketService.updatePromotionInTicket();
+            return ticketService.editTicket(ticketId, ticketRequest);
+        }
+        return null;
+    }
+
+    boolean isAdmin() {
+        String userName = securityUtil.getCurrentUserLogin().orElse(null);
+        if (userName != null && userName.equals("admin")) {
+            return true;
+        }
+        return false;
+    }
+
+    @GetMapping("/test")
+    public String test(Model model) {
+        return "user/test";
     }
 }
